Create informativity_KAJ.py
This script for calculating informativity currently lives in a higher-level folder and integrates a number of printed tests, for the purposes of a course project. The code is ready to be fully integrated, though at this point it only supports one context type - all preceding segments in a word. Preliminary .rst documentation coming shortly..